Seperate fetchData function into DOM versus apiCalls files
