+++ src/apiCalls.js
@@ -4,4 +4,17 @@ const allData = [
     fetch(`https://fitlit-api.herokuapp.com/api/v1/hydration`)
   ]
 
-  export { allData }
+  function fetchData() {
+   return Promise.all(allData)
+      .then((res) => {
+       return Promise.all(res.map((item) => {
+        return item.json();
+        }))
+      })
+      .catch(error => {
+        console.log("error")
+        return error; 
+      })    
+  }
+
+  export { allData, fetchData }
